requirements.txt
pygame==2.6.1
pillow==11.3.0
requests==2.32.5
